requirements.txt
aiofiles==0.6.0
APScheduler==3.7.0
beautifulsoup4==4.9.3
fastapi
lxml==4.6.3
pandas
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.13
uvicorn==0.13.4
xlrd==2.0.1
yahooquery==2.2.15
psycopg2-binary